Add files via upload
Implement login functionality.

Create an array/object of userList with atleast 4 users
On submitting the login form, take the user and password from input fields and check if user exists in the userList array/object
On successful login go to profile page  
On error, show the error detail on HTML page.
Display the details of current logged in user on profile page. Use localStorage to save the current user after login.
If the user is already logged in redirect him directly to profile page on opening login page.
If user is not logged in and he is trying to access the profile page, redirect him to login page.
Clicking on Signout on profie page should clear the localStorage, log out and redirect him to login page.